+++ src/main/java/subway/application/LineService.java
@@ -9,9 +9,9 @@ import subway.controller.dto.LineResponse;
 import subway.controller.dto.SectionCreateRequest;
 import subway.controller.dto.SectionDeleteRequest;
 import subway.dao.entity.SectionEntity;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.line.Section;
+import subway.domain.line.Station;
 import subway.exception.InvalidDistanceException;
 import subway.exception.LineNameException;
 import subway.exception.LineStationAdditionException;
