+++ src/main/java/subway/config/BeanConfig.java
@@ -2,8 +2,8 @@ package subway.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import subway.domain.DistanceFarePolicy;
-import subway.domain.FarePolicy;
+import subway.domain.fare.DistanceFarePolicy;
+import subway.domain.fare.FarePolicy;
 
 @Configuration
 public class BeanConfig {
