+++ src/main/java/subway/domain/fare/FarePolicy.java
@@ -1,4 +1,6 @@
-package subway.domain;
+package subway.domain.fare;
+
+import subway.domain.line.Distance;
 
 public interface FarePolicy {
