+++ src/main/java/subway/domain/line/Line.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import java.util.List;
 import java.util.Objects;
