+++ src/main/java/subway/domain/line/Station.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import java.util.Objects;
 import subway.dao.entity.StationEntity;
