+++ src/main/java/subway/domain/route/JgraphtRouteGraph.java
@@ -1,20 +1,24 @@
-package subway.domain;
+package subway.domain.route;
 
 import java.util.List;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
+import subway.domain.line.Distance;
+import subway.domain.line.Line;
+import subway.domain.line.Section;
+import subway.domain.line.Station;
 
-public class Route {
+public class JgraphtRouteGraph implements RouteGraph {
 
     private final DijkstraShortestPath<Station, DefaultWeightedEdge> sectionGraph;
 
-    private Route(DijkstraShortestPath<Station, DefaultWeightedEdge> sectionGraph) {
+    private JgraphtRouteGraph(DijkstraShortestPath<Station, DefaultWeightedEdge> sectionGraph) {
         this.sectionGraph = sectionGraph;
     }
 
-    public static Route from(List<Line> lines) {
+    public static JgraphtRouteGraph from(List<Line> lines) {
         WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
 
         for (Line line : lines) {
@@ -22,7 +26,7 @@ public class Route {
             addSection(graph, line);
         }
 
-        return new Route(new DijkstraShortestPath<>(graph));
+        return new JgraphtRouteGraph(new DijkstraShortestPath<>(graph));
     }
 
     private static void addStation(WeightedMultigraph<Station, DefaultWeightedEdge> graph, Line line) {
@@ -39,6 +43,7 @@ public class Route {
         }
     }
 
+    @Override
     public List<Station> findShortestRoute(Station start, Station end) {
         GraphPath<Station, DefaultWeightedEdge> path = sectionGraph.getPath(start, end);
 
@@ -48,6 +53,7 @@ public class Route {
         return path.getVertexList();
     }
 
+    @Override
     public Distance findShortestDistance(Station start, Station end) {
         return new Distance((int) sectionGraph.getPathWeight(start, end));
     }
