+++ src/test/java/subway/application/LineServiceTest.java
@@ -20,11 +20,11 @@ import subway.controller.dto.LineRequest;
 import subway.controller.dto.LineResponse;
 import subway.controller.dto.SectionCreateRequest;
 import subway.controller.dto.SectionDeleteRequest;
-import subway.domain.Distance;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Sections;
-import subway.domain.Station;
+import subway.domain.line.Distance;
+import subway.domain.line.Line;
+import subway.domain.line.Section;
+import subway.domain.line.Sections;
+import subway.domain.line.Station;
 import subway.exception.InvalidDistanceException;
 import subway.exception.LineNameException;
 import subway.exception.LineStationAdditionException;
