+++ src/test/java/subway/domain/DistanceTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import subway.domain.line.Distance;
 import subway.exception.InvalidDistanceException;
 
 class DistanceTest {
