+++ src/test/java/subway/domain/JgraphtRouteGraphTest.java
@@ -17,19 +17,23 @@ import static subway.fixture.StationFixture.회현역;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.domain.line.Distance;
+import subway.domain.line.Line;
+import subway.domain.line.Station;
+import subway.domain.route.JgraphtRouteGraph;
 import subway.fixture.LineFixture;
 
 @SuppressWarnings("NonAsciiCharacters")
-class RouteTest {
+class JgraphtRouteGraphTest {
 
     @Test
     @DisplayName("출발역과 도착역 까지 최단 경로를 찾는다.")
     void findShortedRoute1() {
         Station start = 남영역;
         Station end = 선릉역;
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        List<Station> shortestRoute = route.findShortestRoute(start, end);
+        List<Station> shortestRoute = jgraphtRouteGraph.findShortestRoute(start, end);
 
         assertThat(shortestRoute).containsExactly(start, 서울역, 시청역, end);
     }
@@ -39,9 +43,9 @@ class RouteTest {
     void findShortedRoute2() {
         Station start = 사당역;
         Station end = 종각역;
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        List<Station> shortestRoute = route.findShortestRoute(start, end);
+        List<Station> shortestRoute = jgraphtRouteGraph.findShortestRoute(start, end);
 
         assertThat(shortestRoute).containsExactly(start, 용산역, 시청역, end);
     }
@@ -51,9 +55,9 @@ class RouteTest {
     void findShortedDistance() {
         Station start = 남영역;
         Station end = 선릉역;
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        Distance distance = route.findShortestDistance(start, end);
+        Distance distance = jgraphtRouteGraph.findShortestDistance(start, end);
 
         assertThat(distance.getValue()).isEqualTo(20);
     }
@@ -64,9 +68,9 @@ class RouteTest {
         Station start = 역삼역;
         Station end = 선릉역;
 
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        assertThatThrownBy(() -> route.findShortestRoute(start, end))
+        assertThatThrownBy(() -> jgraphtRouteGraph.findShortestRoute(start, end))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -76,9 +80,9 @@ class RouteTest {
         Station start = 선릉역;
         Station end = 회현역;
 
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        assertThatThrownBy(() -> route.findShortestRoute(start, end))
+        assertThatThrownBy(() -> jgraphtRouteGraph.findShortestRoute(start, end))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -88,9 +92,9 @@ class RouteTest {
         Station start = 서초역;
         Station end = 회현역;
 
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        assertThatThrownBy(() -> route.findShortestRoute(start, end))
+        assertThatThrownBy(() -> jgraphtRouteGraph.findShortestRoute(start, end))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -100,19 +104,19 @@ class RouteTest {
         Station start = 신사역;
         Station end = 서울역;
 
-        Route route = createRouteMap();
+        JgraphtRouteGraph jgraphtRouteGraph = createRouteMap();
 
-        assertThatThrownBy(() -> route.findShortestRoute(start, end))
+        assertThatThrownBy(() -> jgraphtRouteGraph.findShortestRoute(start, end))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(start.getName() + "과 " + end.getName() + " 사이의 경로가 존재하지 않습니다.");
     }
 
 
-    private static Route createRouteMap() {
+    private static JgraphtRouteGraph createRouteMap() {
         Line 일호선 = LineFixture.일호선;
         Line 이호선 = LineFixture.이호선;
         Line 삼호선 = LineFixture.삼호선;
         Line 사호선 = LineFixture.사호선;
-        return Route.from(List.of(일호선, 이호선, 삼호선, 사호선));
+        return JgraphtRouteGraph.from(List.of(일호선, 이호선, 삼호선, 사호선));
     }
 }
