+++ src/test/java/subway/domain/StationTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import subway.domain.line.Station;
 import subway.exception.InvalidFormatException;
 
 class StationTest {
